# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/merchant/configuration": {
      "post": {
        "operationId": "MerchantConfigurationController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMerchantConfigurationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Merchant configuration created successfully",
            "headers": {
              "x-tenant-id": {
                "description": "Tenant identifier",
                "schema": {
                  "type": "string"
                }
              },
              "x-api-key": {
                "description": "Current tenant API key",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantConfigurationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create merchant configuration",
        "tags": [
          "Merchant Configuration"
        ]
      },
      "get": {
        "operationId": "MerchantConfigurationController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Merchant configuration retrieved successfully",
            "headers": {
              "x-tenant-id": {
                "description": "Tenant identifier",
                "schema": {
                  "type": "string"
                }
              },
              "x-api-key": {
                "description": "Current tenant API key",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantConfigurationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get merchant configuration",
        "tags": [
          "Merchant Configuration"
        ]
      },
      "patch": {
        "operationId": "MerchantConfigurationController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMerchantConfigurationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Merchant configuration updated successfully",
            "headers": {
              "x-tenant-id": {
                "description": "Tenant identifier",
                "schema": {
                  "type": "string"
                }
              },
              "x-api-key": {
                "description": "Current tenant API key",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantConfigurationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update merchant configuration",
        "tags": [
          "Merchant Configuration"
        ]
      }
    },
    "/api/v1/merchant/configuration/verify/mtn": {
      "post": {
        "operationId": "MerchantConfigurationController_verifyMtn",
        "parameters": [],
        "responses": {
          "200": {
            "description": "MTN credentials verified successfully",
            "headers": {
              "x-tenant-id": {
                "description": "Tenant identifier",
                "schema": {
                  "type": "string"
                }
              },
              "x-api-key": {
                "description": "Current tenant API key",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verify MTN credentials",
        "tags": [
          "Merchant Configuration"
        ]
      }
    },
    "/api/v1/merchant/configuration/verify/airtel": {
      "post": {
        "operationId": "MerchantConfigurationController_verifyAirtel",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Airtel credentials verified successfully",
            "headers": {
              "x-tenant-id": {
                "description": "Tenant identifier",
                "schema": {
                  "type": "string"
                }
              },
              "x-api-key": {
                "description": "Current tenant API key",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verify Airtel credentials",
        "tags": [
          "Merchant Configuration"
        ]
      }
    },
    "/api/v1/merchant/configuration/verify/bank": {
      "post": {
        "operationId": "MerchantConfigurationController_verifyBank",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Bank account verified successfully",
            "headers": {
              "x-tenant-id": {
                "description": "Tenant identifier",
                "schema": {
                  "type": "string"
                }
              },
              "x-api-key": {
                "description": "Current tenant API key",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verify bank account",
        "tags": [
          "Merchant Configuration"
        ]
      }
    },
    "/api/v1/merchant/configuration/webhook/test": {
      "post": {
        "operationId": "MerchantConfigurationController_testWebhook",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestWebhookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook test completed",
            "headers": {
              "x-tenant-id": {
                "description": "Tenant identifier",
                "schema": {
                  "type": "string"
                }
              },
              "x-api-key": {
                "description": "Current tenant API key",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookTestResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Test webhook endpoint",
        "tags": [
          "Merchant Configuration"
        ]
      }
    }
  },
  "info": {
    "title": "Payment Gateway",
    "description": "API documentation for Payment Gateway",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "x-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "x-tenant-id": {
        "type": "apiKey",
        "in": "header",
        "name": "x-tenant-id"
      }
    },
    "schemas": {
...
      "CreateMerchantConfigurationDto": {
        "type": "object",
        "properties": {
          "businessName": {
            "type": "string",
            "description": "Legal business name",
            "example": "Acme Corporation Ltd"
          },
          "businessRegistrationNumber": {
            "type": "string",
            "description": "Business registration number",
            "example": "BRN-123456"
          },
          "taxId": {
            "type": "string",
            "description": "Tax ID / VAT number",
            "example": "TAX-ZM-123456"
          },
          "businessCategory": {
            "type": "string",
            "description": "Business category/industry",
            "example": "E-Commerce"
          },
          "websiteUrl": {
            "type": "string",
            "description": "Business website URL",
            "example": "https://acme.com"
          },
          "businessAddress": {
            "type": "string",
            "description": "Physical business address (will be encrypted)",
            "example": "123 Business Street, Lusaka, ZM"
          },
          "contactPersonName": {
            "type": "string",
            "description": "Primary contact person name",
            "example": "John Doe"
          },
          "contactPersonPhone": {
            "type": "string",
            "description": "Primary contact phone",
            "example": "+260965123456"
          },
          "contactPersonEmail": {
            "type": "string",
            "description": "Primary contact email",
            "example": "contact@acme.com"
          },
          "mtnCollectionSubscriptionKey": {
            "type": "string",
            "description": "MTN collection API subscription key (will be encrypted)",
            "example": "mtncoll_abc123xyz789"
          },
          "mtnCollectionApiKey": {
            "type": "string",
            "description": "MTN collection API key (will be encrypted)",
            "example": "mtnkey_abc123xyz789"
          },
          "mtnCollectionXReferenceId": {
            "type": "string",
            "description": "MTN collection X-Reference-ID",
            "example": "acme-mtn-ref-001"
          },
          "mtnCollectionTargetEnvironment": {
            "type": "string",
            "description": "MTN collection target environment",
            "enum": [
              "sandbox",
              "production"
            ],
            "default": "sandbox"
          },
          "mtnDisbursementSubscriptionKey": {
            "type": "string",
            "description": "MTN disbursement API subscription key (will be encrypted)"
          },
          "mtnDisbursementApiKey": {
            "type": "string",
            "description": "MTN disbursement API key (will be encrypted)"
          },
          "mtnDisbursementXReferenceId": {
            "type": "string",
            "description": "MTN disbursement X-Reference-ID"
          },
          "mtnDisbursementTargetEnvironment": {
            "type": "string",
            "description": "MTN disbursement target environment",
            "enum": [
              "sandbox",
              "production"
            ],
            "default": "sandbox"
          },
          "mtnAccountHolder": {
            "type": "string",
            "description": "MTN account holder name"
          },
          "mtnAccountActive": {
            "type": "boolean",
            "description": "MTN account is active"
          },
          "airtelClientId": {
            "type": "string",
            "description": "Airtel OAuth2 Client ID (will be encrypted)",
            "example": "airtel_client_123"
          },
          "airtelClientSecret": {
            "type": "string",
            "description": "Airtel OAuth2 Client Secret (will be encrypted)"
          },
          "airtelSigningSecret": {
            "type": "string",
            "description": "Airtel Signing Secret (will be encrypted)"
          },
          "airtelEncryptionPublicKey": {
            "type": "string",
            "description": "Airtel RSA Public Key (will be encrypted)"
          },
          "airtelEnvironment": {
            "type": "string",
            "description": "Airtel environment",
            "enum": [
              "staging",
              "production"
            ],
            "default": "staging"
          },
          "airtelCountry": {
            "type": "string",
            "description": "Airtel country code",
            "example": "ZM",
            "default": "ZM"
          },
          "airtelCurrency": {
            "type": "string",
            "description": "Airtel default currency",
            "example": "ZMW",
            "default": "ZMW"
          },
          "airtelMerchantId": {
            "type": "string",
            "description": "Airtel merchant account ID"
          },
          "airtelAccountActive": {
            "type": "boolean",
            "description": "Airtel account is active"
          },
          "bankAccountHolder": {
            "type": "string",
            "description": "Bank account holder name",
            "example": "Acme Corporation"
          },
          "bankAccountNumber": {
            "type": "string",
            "description": "Bank account number (will be encrypted)",
            "example": "1234567890"
          },
          "bankAccountType": {
            "type": "string",
            "description": "Bank account type",
            "enum": [
              "checking",
              "savings",
              "business"
            ]
          },
          "bankName": {
            "type": "string",
            "description": "Bank name",
            "example": "Zambia National Bank"
          },
          "bankBranchCode": {
            "type": "string",
            "description": "Bank branch code",
            "example": "ZNBK-LSKX"
          },
          "bankSwiftCode": {
            "type": "string",
            "description": "SWIFT/BIC code",
            "example": "ZNBKZMLU"
          },
          "bankAccountCurrency": {
            "type": "string",
            "description": "Bank account currency",
            "example": "ZMW",
            "default": "ZMW"
          },
          "kycStatus": {
            "type": "string",
            "description": "KYC verification status",
            "enum": [
              "PENDING",
              "VERIFIED",
              "REJECTED",
              "NEEDS_UPDATE"
            ]
          },
          "directorName": {
            "type": "string",
            "description": "Director/Owner name"
          },
          "directorIdNumber": {
            "type": "string",
            "description": "Director/Owner ID number (will be encrypted)"
          },
          "directorIdType": {
            "type": "string",
            "description": "Director/Owner ID type",
            "enum": [
              "passport",
              "national_id",
              "driver_license"
            ]
          },
          "beneficialOwnerInfo": {
            "type": "string",
            "description": "Beneficial owner information (will be encrypted)"
          },
          "complianceNotes": {
            "type": "string",
            "description": "Compliance notes"
          },
          "webhookUrl": {
            "type": "string",
            "description": "Webhook URL for payment notifications",
            "example": "https://acme.com/webhooks/payments"
          },
          "webhookSecret": {
            "type": "string",
            "description": "Webhook signature key (will be encrypted)"
          },
          "webhookEvents": {
            "description": "Webhook events to subscribe to",
            "example": [
              "payment.success",
              "payment.failed",
              "disbursement.complete"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "webhookEnabled": {
            "type": "boolean",
            "description": "Enable/disable webhook",
            "default": true
          },
          "maxDailyCollections": {
            "type": "number",
            "description": "Maximum daily collection requests",
            "example": 10000,
            "default": 10000
          },
          "maxDailyDisbursementAmount": {
            "type": "number",
            "description": "Maximum daily disbursement amount",
            "example": "100000.00"
          },
          "maxTransactionAmount": {
            "type": "number",
            "description": "Single transaction maximum amount",
            "example": "50000.00"
          },
          "approvalThresholdAmount": {
            "type": "number",
            "description": "Require approval for transactions above this amount",
            "example": "10000.00"
          },
          "notes": {
            "type": "string",
            "description": "Additional notes about merchant configuration"
          }
        },
        "required": [
          "businessName"
        ]
      },
      "MerchantConfigurationResponseDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateMerchantConfigurationDto": {
        "type": "object",
        "properties": {
          "businessName": {
            "type": "string",
            "description": "Legal business name",
            "example": "Acme Corporation Ltd"
          },
          "businessRegistrationNumber": {
            "type": "string",
            "description": "Business registration number",
            "example": "BRN-123456"
          },
          "taxId": {
            "type": "string",
            "description": "Tax ID / VAT number",
            "example": "TAX-ZM-123456"
          },
          "businessCategory": {
            "type": "string",
            "description": "Business category/industry",
            "example": "E-Commerce"
          },
          "websiteUrl": {
            "type": "string",
            "description": "Business website URL",
            "example": "https://acme.com"
          },
          "businessAddress": {
            "type": "string",
            "description": "Physical business address (will be encrypted)",
            "example": "123 Business Street, Lusaka, ZM"
          },
          "contactPersonName": {
            "type": "string",
            "description": "Primary contact person name",
            "example": "John Doe"
          },
          "contactPersonPhone": {
            "type": "string",
            "description": "Primary contact phone",
            "example": "+260965123456"
          },
          "contactPersonEmail": {
            "type": "string",
            "description": "Primary contact email",
            "example": "contact@acme.com"
          },
          "mtnCollectionSubscriptionKey": {
            "type": "string",
            "description": "MTN collection API subscription key (will be encrypted)",
            "example": "mtncoll_abc123xyz789"
          },
          "mtnCollectionApiKey": {
            "type": "string",
            "description": "MTN collection API key (will be encrypted)",
            "example": "mtnkey_abc123xyz789"
          },
          "mtnCollectionXReferenceId": {
            "type": "string",
            "description": "MTN collection X-Reference-ID",
            "example": "acme-mtn-ref-001"
          },
          "mtnCollectionTargetEnvironment": {
            "type": "string",
            "description": "MTN collection target environment",
            "enum": [
              "sandbox",
              "production"
            ],
            "default": "sandbox"
          },
          "mtnDisbursementSubscriptionKey": {
            "type": "string",
            "description": "MTN disbursement API subscription key (will be encrypted)"
          },
          "mtnDisbursementApiKey": {
            "type": "string",
            "description": "MTN disbursement API key (will be encrypted)"
          },
          "mtnDisbursementXReferenceId": {
            "type": "string",
            "description": "MTN disbursement X-Reference-ID"
          },
          "mtnDisbursementTargetEnvironment": {
            "type": "string",
            "description": "MTN disbursement target environment",
            "enum": [
              "sandbox",
              "production"
            ],
            "default": "sandbox"
          },
          "mtnAccountHolder": {
            "type": "string",
            "description": "MTN account holder name"
          },
          "mtnAccountActive": {
            "type": "boolean",
            "description": "MTN account is active"
          },
          "airtelClientId": {
            "type": "string",
            "description": "Airtel OAuth2 Client ID (will be encrypted)",
            "example": "airtel_client_123"
          },
          "airtelClientSecret": {
            "type": "string",
            "description": "Airtel OAuth2 Client Secret (will be encrypted)"
          },
          "airtelSigningSecret": {
            "type": "string",
            "description": "Airtel Signing Secret (will be encrypted)"
          },
          "airtelEncryptionPublicKey": {
            "type": "string",
            "description": "Airtel RSA Public Key (will be encrypted)"
          },
          "airtelEnvironment": {
            "type": "string",
            "description": "Airtel environment",
            "enum": [
              "staging",
              "production"
            ],
            "default": "staging"
          },
          "airtelCountry": {
            "type": "string",
            "description": "Airtel country code",
            "example": "ZM",
            "default": "ZM"
          },
          "airtelCurrency": {
            "type": "string",
            "description": "Airtel default currency",
            "example": "ZMW",
            "default": "ZMW"
          },
          "airtelMerchantId": {
            "type": "string",
            "description": "Airtel merchant account ID"
          },
          "airtelAccountActive": {
            "type": "boolean",
            "description": "Airtel account is active"
          },
          "bankAccountHolder": {
            "type": "string",
            "description": "Bank account holder name",
            "example": "Acme Corporation"
          },
          "bankAccountNumber": {
            "type": "string",
            "description": "Bank account number (will be encrypted)",
            "example": "1234567890"
          },
          "bankAccountType": {
            "type": "string",
            "description": "Bank account type",
            "enum": [
              "checking",
              "savings",
              "business"
            ]
          },
          "bankName": {
            "type": "string",
            "description": "Bank name",
            "example": "Zambia National Bank"
          },
          "bankBranchCode": {
            "type": "string",
            "description": "Bank branch code",
            "example": "ZNBK-LSKX"
          },
          "bankSwiftCode": {
            "type": "string",
            "description": "SWIFT/BIC code",
            "example": "ZNBKZMLU"
          },
          "bankAccountCurrency": {
            "type": "string",
            "description": "Bank account currency",
            "example": "ZMW",
            "default": "ZMW"
          },
          "kycStatus": {
            "type": "string",
            "description": "KYC verification status",
            "enum": [
              "PENDING",
              "VERIFIED",
              "REJECTED",
              "NEEDS_UPDATE"
            ]
          },
          "directorName": {
            "type": "string",
            "description": "Director/Owner name"
          },
          "directorIdNumber": {
            "type": "string",
            "description": "Director/Owner ID number (will be encrypted)"
          },
          "directorIdType": {
            "type": "string",
            "description": "Director/Owner ID type",
            "enum": [
              "passport",
              "national_id",
              "driver_license"
            ]
          },
          "beneficialOwnerInfo": {
            "type": "string",
            "description": "Beneficial owner information (will be encrypted)"
          },
          "complianceNotes": {
            "type": "string",
            "description": "Compliance notes"
          },
          "webhookUrl": {
            "type": "string",
            "description": "Webhook URL for payment notifications",
            "example": "https://acme.com/webhooks/payments"
          },
          "webhookSecret": {
            "type": "string",
            "description": "Webhook signature key (will be encrypted)"
          },
          "webhookEvents": {
            "description": "Webhook events to subscribe to",
            "example": [
              "payment.success",
              "payment.failed",
              "disbursement.complete"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "webhookEnabled": {
            "type": "boolean",
            "description": "Enable/disable webhook",
            "default": true
          },
          "maxDailyCollections": {
            "type": "number",
            "description": "Maximum daily collection requests",
            "example": 10000,
            "default": 10000
          },
          "maxDailyDisbursementAmount": {
            "type": "number",
            "description": "Maximum daily disbursement amount",
            "example": "100000.00"
          },
          "maxTransactionAmount": {
            "type": "number",
            "description": "Single transaction maximum amount",
            "example": "50000.00"
          },
          "approvalThresholdAmount": {
            "type": "number",
            "description": "Require approval for transactions above this amount",
            "example": "10000.00"
          },
          "notes": {
            "type": "string",
            "description": "Additional notes about merchant configuration"
          },
          "isActive": {
            "type": "boolean",
            "description": "Activate/deactivate merchant configuration"
          }
        },
        "required": [
          "businessName"
        ]
      },
      "VerificationResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether verification was successful",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Verification result message",
            "example": "MTN credentials verified successfully"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
      "TestWebhookDto": {
        "type": "object",
        "properties": {
          "eventType": {
            "type": "string",
            "description": "Webhook event type to test",
            "example": "payment.success"
          },
          "payload": {
            "type": "object",
            "description": "Test payload to send"
          }
        },
        "required": [
          "eventType"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    },
    {
      "x-api-key": []
    },
    {
      "x-tenant-id": []
    }
  ]
}